Fix package format for pip
Having the README_PIP.md and other files in the top level of the package
meant that they aren't included when `python setup.py bdist_egg` is
run.

(This is so that they don't end up in the top-level of the
./site-packages/ folder when the package is installed.)

This commit:

1. Moves README_PIP.md to the crispy_forms_semantic_ui/ folder (and
   renames it README).

   Also took the liberty of re-writing the README_PIP.md file as rst :)

2. Modifies setup.py to locate the README file for the long description
   in its new location

   (This is a bit hacky due to the fact that the app name uses "-" whilst
   the folder name uses "_". You might want to change this? :) )

When built using bdist and bdist_egg all required files are included, so
the package should now be able to be downloaded and installed using pip.

Signed-off-by: Rod Manning <rod.t.manning@gmail.com>